- remove idle job-pausing feature from instrumentation - extract websocket server setup into dedicated module
