+++ apps/nextjs/src/instrumentation.ts
@@ -17,138 +17,8 @@ export async function register() {
     await jobGroup.initializeAsync();
     await jobGroup.startAllAsync();
 
-    // Start WebSocket server in the same process (eliminates separate wssServer process)
-    const { WebSocketServer } = await import("ws");
-    const { applyWSSHandler } = await import("@trpc/server/adapters/ws");
-    const { appRouter, createTRPCContext } = await import("@homarr/api/websocket");
-    const { getSessionFromToken, sessionTokenCookieName } = await import("@homarr/auth");
-    const { parseCookies } = await import("@homarr/common");
-    const { db } = await import("@homarr/db");
-
-    const wss = new WebSocketServer({ port: 3001 });
-
-    applyWSSHandler({
-      wss,
-      router: appRouter,
-      // eslint-disable-next-line no-restricted-syntax
-      createContext: async ({ req }) => {
-        try {
-          const headers = Object.entries(req.headers).map(
-            ([key, value]) => [key, typeof value === "string" ? value : value?.[0]] as [string, string],
-          );
-          const nextHeaders = new Headers(headers);
-          const store = parseCookies(nextHeaders.get("cookie") ?? "");
-          const sessionToken = store[sessionTokenCookieName];
-          const session = await getSessionFromToken(db, sessionToken);
-          return createTRPCContext({ headers: nextHeaders, session });
-        } catch (error) {
-          logger.error(error);
-          return createTRPCContext({ headers: new Headers(), session: null });
-        }
-      },
-      keepAlive: {
-        enabled: true,
-        pingMs: 30000,
-        pongWaitMs: 5000,
-      },
-    });
-
-    // Idle detection: pause integration cron jobs when no clients are connected
-    const IDLE_GRACE_MS = 60_000;
-    type JobKey = Parameters<typeof jobGroup.stopAsync>[0];
-    const INTEGRATION_JOB_KEYS: JobKey[] = [
-      "ping",
-      "smartHomeEntityState",
-      "mediaServer",
-      "mediaOrganizer",
-      "downloads",
-      "dnsHole",
-      "mediaRequestStats",
-      "mediaRequestList",
-      "rssFeeds",
-      "indexerManager",
-      "healthMonitoring",
-      "mediaTranscoding",
-      "minecraftServerStatus",
-      "dockerContainers",
-      "networkController",
-      "firewallCpu",
-      "firewallMemory",
-      "firewallVersion",
-      "firewallInterfaces",
-      "refreshNotifications",
-      "weather",
-      "timetable",
-      "tracearr",
-    ];
-
-    let idleTimer: ReturnType<typeof setTimeout> | null = null;
-    let isIdle = false;
-
-    const pauseIntegrationJobsAsync = async () => {
-      if (isIdle) return;
-      isIdle = true;
-      logger.info("No active clients - pausing integration cron jobs");
-      for (const name of INTEGRATION_JOB_KEYS) {
-        await jobGroup.stopAsync(name);
-      }
-    };
-
-    const resumeIntegrationJobsAsync = async () => {
-      if (!isIdle) return;
-      isIdle = false;
-      logger.info("Client connected - resuming integration cron jobs");
-      for (const name of INTEGRATION_JOB_KEYS) {
-        await jobGroup.startAsync(name);
-      }
-    };
-
-    // Start idle timer immediately - pause jobs if no client connects within grace period
-    idleTimer = setTimeout(() => {
-      idleTimer = null;
-      void pauseIntegrationJobsAsync();
-    }, IDLE_GRACE_MS);
-
-    let externalClients = 0;
-
-    const isLocalhost = (address: string | undefined) =>
-      address === "127.0.0.1" || address === "::1" || address === "::ffff:127.0.0.1";
-
-    wss.on("connection", (websocket, incomingMessage) => {
-      const remote = incomingMessage.socket.remoteAddress;
-      const external = !isLocalhost(remote);
-      if (external) externalClients++;
-      logger.info(
-        `Connection (external=${externalClients}) ${remote} ${incomingMessage.method} ${incomingMessage.url}`,
-      );
-
-      if (external) {
-        if (idleTimer !== null) {
-          clearTimeout(idleTimer);
-          idleTimer = null;
-        }
-        void resumeIntegrationJobsAsync();
-      }
-
-      websocket.once("close", (code, reason) => {
-        if (external) externalClients--;
-        logger.info(`Disconnected (external=${externalClients}) ${code} ${reason.toString()}`);
-        if (external && externalClients === 0) {
-          idleTimer = setTimeout(() => {
-            idleTimer = null;
-            void pauseIntegrationJobsAsync();
-          }, IDLE_GRACE_MS);
-        }
-      });
-    });
-
-    logger.info("WebSocket server listening on ws://localhost:3001");
-
-    process.on("SIGTERM", () => {
-      if (idleTimer !== null) clearTimeout(idleTimer);
-      wss.clients.forEach((ws) => ws.close());
-      wss.close();
-    });
+    const { initWebSocketServerAsync } = await import("./server/websocket");
+    await initWebSocketServerAsync();
 
   }
 }
